#2049 Moved more common functionality into base class.
